Simplified analytics utils
