+++ app/sync/session.js
@@ -1,7 +1,7 @@
 import srp from 'srp-js';
 import * as crypt from './crypt';
 import * as util from '../common/fetch';
-import {trackEvent, setAccountId} from '../common/analytics';
+import {trackEvent} from '../common/analytics';
 
 /** Create a new session for the user */
 export async function login (rawEmail, rawPassphrase) {
@@ -77,7 +77,6 @@ export async function login (rawEmail, rawPassphrase) {
   );
 
   // Set the ID for Google Analytics
-  setAccountId(accountId);
   trackEvent('Session', 'Login');
 }
