+++ app/ui/components/base/link.js
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react';
 import autobind from 'autobind-decorator';
-import {trackEvent} from '../../../common/analytics/index';
+import {trackEvent} from '../../../common/analytics';
 import * as misc from '../../../common/misc';
 
 type Props = {|
