+++ app/ui/components/dropdowns/content-type-dropdown.js
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import autobind from 'autobind-decorator';
 import {Dropdown, DropdownButton, DropdownDivider, DropdownItem} from '../base/dropdown';
-import {trackEvent} from '../../../common/analytics/index';
+import {trackEvent} from '../../../common/analytics';
 import {CONTENT_TYPE_FILE, CONTENT_TYPE_FORM_DATA, CONTENT_TYPE_FORM_URLENCODED, CONTENT_TYPE_GRAPHQL, CONTENT_TYPE_JSON, CONTENT_TYPE_OTHER, CONTENT_TYPE_XML, getContentTypeName} from '../../../common/constants';
 import {showModal} from '../modals/index';
 import AlertModal from '../modals/alert-modal';
