+++ app/ui/components/dropdowns/method-dropdown.js
@@ -4,7 +4,7 @@ import autobind from 'autobind-decorator';
 import {Dropdown, DropdownButton, DropdownDivider, DropdownItem} from '../base/dropdown';
 import * as constants from '../../../common/constants';
 import {showPrompt} from '../modals/index';
-import {trackEvent} from '../../../common/analytics/index';
+import {trackEvent} from '../../../common/analytics';
 
 const LOCALSTORAGE_KEY = 'insomnia.httpMethods';
