+++ app/ui/components/editors/auth/aws-auth.js
@@ -2,7 +2,7 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import autobind from 'autobind-decorator';
 import KeyValueEditor from '../../key-value-editor/editor';
-import {trackEvent} from '../../../../common/analytics/index';
+import {trackEvent} from '../../../../common/analytics';
 import {AUTH_AWS_IAM} from '../../../../common/constants';
 
 @autobind
