+++ app/ui/components/editors/auth/ntlm-auth.js
@@ -2,7 +2,7 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import autobind from 'autobind-decorator';
 import KeyValueEditor from '../../key-value-editor/editor';
-import {trackEvent} from '../../../../common/analytics/index';
+import {trackEvent} from '../../../../common/analytics';
 import {AUTH_NTLM} from '../../../../common/constants';
 
 @autobind
