+++ app/ui/components/editors/auth/o-auth-2-auth.js
@@ -13,7 +13,7 @@ import accessTokenUrls from '../../../../datasets/access-token-urls';
 import getAccessToken from '../../../../network/o-auth-2/get-token';
 import * as models from '../../../../models';
 import Link from '../../base/link';
-import {trackEvent} from '../../../../common/analytics/index';
+import {trackEvent} from '../../../../common/analytics';
 import HelpTooltip from '../../help-tooltip';
 import PromptButton from '../../base/prompt-button';
 import TimeFromNow from '../../time-from-now';
