+++ app/ui/components/editors/body/file-editor.js
@@ -6,7 +6,7 @@ import electron from 'electron';
 import FileInputButton from '../../base/file-input-button';
 import PromptButton from '../../base/prompt-button';
 import * as misc from '../../../../common/misc';
-import {trackEvent} from '../../../../common/analytics/index';
+import {trackEvent} from '../../../../common/analytics';
 
 @autobind
 class FileEditor extends PureComponent {
