+++ app/ui/components/editors/body/url-encoded-editor.js
@@ -2,7 +2,7 @@ import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import autobind from 'autobind-decorator';
 import KeyValueEditor from '../../key-value-editor/editor';
-import {trackEvent} from '../../../../common/analytics/index';
+import {trackEvent} from '../../../../common/analytics';
 
 @autobind
 class UrlEncodedEditor extends PureComponent {
