+++ app/ui/components/settings/plugins.js
@@ -6,7 +6,7 @@ import autobind from 'autobind-decorator';
 import * as electron from 'electron';
 import Button from '../base/button';
 import CopyButton from '../base/copy-button';
-import {trackEvent} from '../../../common/analytics/index';
+import {trackEvent} from '../../../common/analytics';
 import {reload} from '../../../templating/index';
 import installPlugin from '../../../plugins/install';
 import HelpTooltip from '../help-tooltip';
